+++ milkman-rest/src/main/java/milkman/ui/plugin/rest/domain/RestRequestContainer.java
@@ -33,7 +33,7 @@ public class RestRequestContainer extends RequestContainer {
 
 	@Override
 	public String getType() {
-		return httpMethod;
+		return "HTTP";
 	}
 
 	@Override